+++ packages/core/src/components/lf-code/prism.tsx.highlight.ts
@@ -1,13 +1,13 @@
 // @ts-nocheck
 export const LF_CODE_TSX = (Prism: Prism.Environment) => {
   const a = Prism.util.clone(Prism.languages.typescript);
-  (Prism.languages.tsx = Prism.languages.extend("jsx", a)),
+  ((Prism.languages.tsx = Prism.languages.extend("jsx", a)),
     delete Prism?.languages?.tsx?.parameter,
-    delete Prism?.languages?.tsx?.["literal-property"];
+    delete Prism?.languages?.tsx?.["literal-property"]);
   const t = Prism.languages?.tsx?.tag;
-  (t.pattern = RegExp(
+  ((t.pattern = RegExp(
     "(^|[^\\w$]|(?=</))(?:" + t.pattern.source + ")",
     t.pattern.flags,
   )),
-    (t.lookbehind = !0);
+    (t.lookbehind = !0));
 };