+++ packages/assets/CHANGELOG.md
@@ -2,6 +2,30 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.4.0-rc.3](https://github.com/lucafoscili/lf-assets/compare/0.4.0-rc.2...0.4.0-rc.3) (2025-08-29)
+
+**Note:** Version bump only for package @lf-widgets/assets
+
+# [0.4.0-rc.2](https://github.com/lucafoscili/lf-assets/compare/0.4.0-rc.1...0.4.0-rc.2) (2025-08-29)
+
+**Note:** Version bump only for package @lf-widgets/assets
+
+# [0.4.0-rc.1](https://github.com/lucafoscili/lf-assets/compare/0.4.0-rc.0...0.4.0-rc.1) (2025-08-28)
+
+### Features
+
+- Add LLM mock toggler to showcase component ([52e4b8d](https://github.com/lucafoscili/lf-assets/commit/52e4b8d1fa33fb2bc3f73482ed15710992c66bbe))
+- Add sprite config and improve SVG sprite handling ([fbd46ce](https://github.com/lucafoscili/lf-assets/commit/fbd46ce3e0faaa9a3ff4f10b7ce01bc14bf40cfa))
+- Add SVG sprite support to lf-image component ([e300e7d](https://github.com/lucafoscili/lf-assets/commit/e300e7dbe75a0fefe60b00d638fa080c1ec4b06f))
+
+# [0.4.0-rc.0](https://github.com/lucafoscili/lf-assets/compare/0.3.8-rc.0...0.4.0-rc.0) (2025-08-27)
+
+**Note:** Version bump only for package @lf-widgets/assets
+
+## [0.3.8-rc.0](https://github.com/lucafoscili/lf-assets/compare/0.3.7...0.3.8-rc.0) (2025-08-22)
+
+**Note:** Version bump only for package @lf-widgets/assets
+
 ## [0.3.7](https://github.com/lucafoscili/lf-assets/compare/0.3.7-rc.0...0.3.7) (2025-05-08)
 
 **Note:** Version bump only for package @lf-widgets/assets